+++ Code/Sensor_main/src/LoraMeteo.h
@@ -15,11 +15,9 @@ struct tDataPacket
 {
   float Temp;
   float Hum;
-  int Pres;
-  int BatteryLevelRaw;
+  int32_t Pres;
+  int8_t BatteryLevelRaw;
 };
-
-
 
 int InitLoraServer(int CsPin, int ResetPin, int LoraSpreadingFactor);
